requirements.txt
streamlit>=1.30.0
yfinance>=0.2.30
pandas>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
